backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
websockets==13.1
python-multipart==0.0.18
pydantic==2.9.2
azure-cognitiveservices-speech==1.41.1
openai==1.55.3
anthropic==0.39.0
jellyfish==1.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
numpy==2.1.3
aiofiles==24.1.0
python-dotenv==1.0.1
